# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  name: string;
  location: string;
  price: string;
  bedrooms: number;
  bathrooms: number;
  area_sqft: number;
  description: string;
  /** @nullable */
  image_url: string | null;
  is_sold_out: boolean;
  is_featured: boolean;
  property_type: string;
  created_at: string;
  updated_at: string;
}

export interface CreatePropertyBody {
  name: string;
  location: string;
  price: string;
  bedrooms: number;
  bathrooms: number;
  area_sqft: number;
  description: string;
  /** @nullable */
  image_url?: string | null;
  is_sold_out?: boolean;
  is_featured?: boolean;
  property_type: string;
}

export interface UpdatePropertyBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  area_sqft?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image_url?: string | null;
  /** @nullable */
  is_sold_out?: boolean | null;
  /** @nullable */
  is_featured?: boolean | null;
  /** @nullable */
  property_type?: string | null;
}

export interface PropertyStats {
  total: number;
  available: number;
  sold_out: number;
  featured: number;
}

export interface SpecialOffer {
  id: number;
  text: string;
  updated_at: string;
}

export interface UpdateSpecialOfferBody {
  text: string;
}

export type ListPropertiesParams = {
  /**
   * @nullable
   */
  sold_out?: boolean | null;
};