# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all properties
 */
export const ListPropertiesQueryParams = zod.object({
  sold_out: zod.coerce.boolean().nullish(),
});

export const ListPropertiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  price: zod.string(),
  bedrooms: zod.number(),
  bathrooms: zod.number(),
  area_sqft: zod.number(),
  description: zod.string(),
  image_url: zod.string().nullable(),
  is_sold_out: zod.boolean(),
  is_featured: zod.boolean(),
  property_type: zod.string(),
  created_at: zod.string(),
  updated_at: zod.string(),
});
export const ListPropertiesResponse = zod.array(ListPropertiesResponseItem);

/**
 * @summary Create a new property
 */
export const CreatePropertyBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  price: zod.string(),
  bedrooms: zod.number(),
  bathrooms: zod.number(),
  area_sqft: zod.number(),
  description: zod.string(),
  image_url: zod.string().nullish(),
  is_sold_out: zod.boolean().optional(),
  is_featured: zod.boolean().optional(),
  property_type: zod.string(),
});

/**
 * @summary Get featured (non-sold-out) properties for homepage
 */
export const GetFeaturedPropertiesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  price: zod.string(),
  bedrooms: zod.number(),
  bathrooms: zod.number(),
  area_sqft: zod.number(),
  description: zod.string(),
  image_url: zod.string().nullable(),
  is_sold_out: zod.boolean(),
  is_featured: zod.boolean(),
  property_type: zod.string(),
  created_at: zod.string(),
  updated_at: zod.string(),
});
export const GetFeaturedPropertiesResponse = zod.array(
  GetFeaturedPropertiesResponseItem,
);

/**
 * @summary Get aggregate stats for admin dashboard
 */
export const GetPropertyStatsResponse = zod.object({
  total: zod.number(),
  available: zod.number(),
  sold_out: zod.number(),
  featured: zod.number(),
});

/**
 * @summary Get a single property
 */
export const GetPropertyParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPropertyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  price: zod.string(),
  bedrooms: zod.number(),
  bathrooms: zod.number(),
  area_sqft: zod.number(),
  description: zod.string(),
  image_url: zod.string().nullable(),
  is_sold_out: zod.boolean(),
  is_featured: zod.boolean(),
  property_type: zod.string(),
  created_at: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Update a property
 */
export const UpdatePropertyParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePropertyBody = zod.object({
  name: zod.string().nullish(),
  location: zod.string().nullish(),
  price: zod.string().nullish(),
  bedrooms: zod.number().nullish(),
  bathrooms: zod.number().nullish(),
  area_sqft: zod.number().nullish(),
  description: zod.string().nullish(),
  image_url: zod.string().nullish(),
  is_sold_out: zod.boolean().nullish(),
  is_featured: zod.boolean().nullish(),
  property_type: zod.string().nullish(),
});

export const UpdatePropertyResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  price: zod.string(),
  bedrooms: zod.number(),
  bathrooms: zod.number(),
  area_sqft: zod.number(),
  description: zod.string(),
  image_url: zod.string().nullable(),
  is_sold_out: zod.boolean(),
  is_featured: zod.boolean(),
  property_type: zod.string(),
  created_at: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Delete a property
 */
export const DeletePropertyParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get the homepage marquee/banner special offer text
 */
export const GetSpecialOfferResponse = zod.object({
  id: zod.number(),
  text: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Update the homepage special offer banner
 */
export const UpdateSpecialOfferBody = zod.object({
  text: zod.string(),
});

export const UpdateSpecialOfferResponse = zod.object({
  id: zod.number(),
  text: zod.string(),
  updated_at: zod.string(),
});