# lib/api-zod/src/generated/types/createPropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePropertyBody {
  name: string;
  location: string;
  price: string;
  bedrooms: number;
  bathrooms: number;
  area_sqft: number;
  description: string;
  /** @nullable */
  image_url?: string | null;
  is_sold_out?: boolean;
  is_featured?: boolean;
  property_type: string;
}