# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  name: string;
  location: string;
  price: string;
  bedrooms: number;
  bathrooms: number;
  area_sqft: number;
  description: string;
  /** @nullable */
  image_url: string | null;
  is_sold_out: boolean;
  is_featured: boolean;
  property_type: string;
  created_at: string;
  updated_at: string;
}