# lib/api-zod/src/generated/types/propertyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyStats {
  total: number;
  available: number;
  sold_out: number;
  featured: number;
}