# lib/api-zod/src/generated/types/specialOffer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface SpecialOffer {
  id: number;
  text: string;
  updated_at: string;
}