# lib/api-zod/src/generated/types/updatePropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Luxury Real Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePropertyBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  bedrooms?: number | null;
  /** @nullable */
  bathrooms?: number | null;
  /** @nullable */
  area_sqft?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image_url?: string | null;
  /** @nullable */
  is_sold_out?: boolean | null;
  /** @nullable */
  is_featured?: boolean | null;
  /** @nullable */
  property_type?: string | null;
}